linux-esx: IPVS modules, built-in serial driver
Docker Swarm requires IP VS features enabled, so enable them as a modules.

VIC team requested serial driver to be built-in instead of module, but
serial driver initialization adds +26ms to the boot time. I figured out
that most of the time (25ms) it tries to detect sizes of fifo buffers to
detect "U6 16550A". But in our hypervisor we emulate only 16550A serial
device. Patch is added to skip this probbing.

Change-Id: Icead05fe88dfc72531f0a122b217ed2da47e21e5
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/1660
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Sharath George