linux-aws: Add enhancements from Amazon
While at it, remove the out-of-tree ENA driver from the spec file, as
this driver is included in the aws patches.

Unrelated fix:

The incremental build logic was supposed to trigger a full-rebuild
whenever something under the support/ directory is modified; instead
it was doing it if any file had the word "support" in it (eg: some of
the linux-aws patches in this change). Thanks to Alexey for the
Makefile fix, which is included here.

Change-Id: Ieb1babc35f4f43d9cd5ef6e653b2085dfde6e3b2
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/5910
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>