linux-esx: remove OOM_SCORE_ADJ_MAX limit check
This change adds a patch to the linux-esx build that removes the
OOM_SCORE_ADJ_MAX limit check. The background as to why we would like to
see this integrated is as follows: We would like to have a canary
process running inside a Photon based VM, with the intention of it being
killed first by Linux when memory pressure is high. Upon encountering
such an OOM kill we can then cleanly shut down the VM.
Unfortunately, vanilla Linux and the hints it arms us with
(oom_score_adj most importantly) was found to be insufficient: the
kernel would still kill processes that actually cause the high memory
pressure first, instead of giving preference to our canary process. With
the patch added as part of this change we effectively allow for setting
a higher priority for the canary process, causing the canary to truly be
the first target. Because we only increase the number of available
priorities, existing workloads are not affected negatively.

Change-Id: I8be7e9f16eb71be684c7eb63482c19f90782866e
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/7768
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>