kernels: Add enhances dependency for hmacgen subpackage
This change is to address the issue cuased while installing
photon-checksum-generator package. Photon-checksum-generator has
"or" of multiple packages (linux-hmacgen) in Requires. while installing
this package, libsolv brings in linux-aws and linux-aws-hmacgen packages
as dependencies even though linux generic was already installed on the machine.
This issue happens because libsolv is trying to resolve the "or" of
Requires dependency in alphabetical order without checking what is installed
on the system. Upstream libsolv community suggested to use "Enhances:" in
each hmacgen subpackage to overcome this issue as solver while resolving the
dependency does not know what linux flavor is installed on the machine.

Change-Id: I35da4f816fab6e1ee96a04d31822b98a13b6356d
Signed-off-by: Keerthana K <keerthanak@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/9255
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Srinidhi Rao <srinidhir@vmware.com>
Reviewed-by: Priyesh Padmavilasom <ppadmavilasom@vmware.com>