kernels: Fix rsa tests in self tests
In the earlier commit the test vectors for rsa 3072, 4098
were taken from pkcs1pad_rsa_tv_template. A message
with the same rsa key with and without padding produced
a different cipher text, casuing the tests to fail.

Updated the message, cipher text that match when
rsa is used without pkcs1pad.

Change-Id: I19155fd1af157576804028e8c36265a4289d16a5
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/22383
Reviewed-by: Keerthana K <keerthanak@vmware.com>
Tested-by: Keerthana K <keerthanak@vmware.com>